+++ NEWS.md
@@ -4,6 +4,8 @@ ClimaCalibrate.jl Release Notes
 main
 -------
 
+- Add `residual` for computing the normalized residual `(mean(G) - obs) / σ`
+  [#350](https://github.com/CliMA/ClimaCalibrate.jl/issues/350)
 - Fix a bug where `SampleBuilder.build_samples` did not work with `OutputVar`s
   with no dimensions. `SampleBuilder`, `ObservationRecipe`, and
   `EnsembleBuilder` now support `OutputVar`s with no dimensions. This requires